move get-session to global lib
