+++ frontend/src/app/dashboard/page.tsx
@@ -1,13 +1,13 @@
 import { Metadata } from "next";
 
 import ClientPage from "@/app/dashboard/page-client";
-import { getSession } from "@/features/account/get-session";
 import { ROUTES } from "@/lib/utils/api/endpoints";
 import { ApiErrorResponse } from "@/lib/utils/api/fetch-wrapper";
 import handleErrorResponse from "@/lib/utils/api/handle-api-error";
 import { processDashboardData } from "@/lib/utils/api/processors/process-dashboard-data";
 import { serverGet } from "@/lib/utils/api/server-fetch";
 import { constructMetadata } from "@/lib/utils/construct-metadata";
+import { getSession } from "@/lib/utils/get-session";
 
 // Explicitly set this page to be dynamic so Next.js doesn't try to statically build it
 export const dynamic = "force-dynamic";
